refactor: Complete backend-agnostic naming cleanup
File renames:
- askGeminiSessionManager.ts → askSessionManager.ts

Category updates:
- Changed tool category from 'gemini' to 'ai' in registry and all tools

Status message updates:
- Removed hardcoded "Gemini" from progress messages in index.ts
- Updated SANDBOX_EXECUTING and PROCESSING_CONTINUE in constants.ts
- Removed unused GEMINI_RESPONSE constant

Parameter/variable renames:
- geminiResponse → responseText in reviewResponseParser.ts
- geminiResponse → responseText in changeModeParser.ts

Comment updates:
- Updated JSDoc comments in reviewResponseParser.ts, reviewPromptBuilder.ts
- Updated function comment in brainstorm.tool.ts parseIdeasFromResponse
